Speed smoothing added
Added speed smoothing to the script to help combat large errors from popping up in the output feed. Ex. a car is traveling at a steady 30 KMH, but one frame reads it as 55 KMH, this 55 KMH will show up in the output video if smoothing is disabled, but if smoothing is enabled, then there will be a much smaller jump in speed spread out over a longer period of time.

I hope that makes sense: if it doesn't, then you can just keep the smoothing feature disabled.