Fixed YOLO and tried to fix 3D bounding boxes
Fixed a bug in the YOLO detection code that prevented it from working - YOLO detection now works. The code between YOLO and Mask R-CNN still needs to be cleaned up, there's a lot of shared code between the two that needs to be aggregated to just one source.

Tweaked the 3D bounding box code to try and make it fit the vehicles better. It's still not great, but it's getting better? Code is dirty and needs to be cleaned up.