Fixed homography saving and added MPH option
Homography matrices need to be saved along with their scaling factor so that the math all checks out if you load someone else's homography file - before this commit, the homography scaling factor was not saved, now it is. So, both the matrix and the scaling factor are saved and loaded together now.

Also, an option to convert from KMH to MPH was added to make life a little easier for the user.