Fixed mask scaling
Vehicle masks (from Mask R-CNN) now scale correctly with the bounding boxes, and can be drawn even if the mask spills off the edge of the screen.

Fixed an odd bug where Mask R-CNN was detecting a COCO object that we didn't have a label for? We only need vehicles, so I just added in a small patch to fix it and moved on.