Fixed overlapping mask graphical bug
Previously, when two masks were being drawn in the same space, the last mask to be drawn would take precedence, and draw its solid black background on top of the other mask, resulting in a graphical bug where a vehicle would look like its mask was being eaten away. Now, all masks have transparent backgrounds for drawing purposes, and can merge cleanly with other, nearby masks.

Out of curiosity, I timed the original method for drawing masks and the new method. For the same video and config settings, the original method took around 0.02 seconds while the new method took around 0.04 seconds, a 100% increase. However, compared to the time it takes for tracking (~0.7 seconds on my CPU for CSRT) and detection (2 or 3 seconds on my CPU for Mask R-CNN), this is relatively cheap.