Parked Vehicle Detection
The script can now detect if a vehicle is parked (i.e. the vehicle has remained under a certain threshold of speed for a certain amount of time). If it is, then the vehicle's speed will be reported as 0. I plan on adding a clearer indication that the vehicle is parked later on.