Pixel to Meter Ratio Estimation with Angle of Motion
The pixel-to-meter ratio is now estimated using a novel method, where we take the angle of motion, the perpendicular angle of motion (pointing towards the ground), we find the lowest point on the vehicle, and then we create a line through that point with the angle of motion and find where it intersects with the bounding box, and we take those two intersections and plot their distance on the top-down image.

The results from this version of the code are very good.